Created NSFV base class
Refs #23794
